refactor more ui components
