+++ src/features/dashboard/super/admin-permissions/molecules/PermissionCard.tsx
@@ -1,14 +1,13 @@
 "use client";
 import React from "react";
 import { useTooltip } from "@/shared/hooks";
-import { Icon } from "@/tw-styled/ui";
 import { useIcons } from "@/tw-styled/tools";
 import {
   useIsPermissionEnabled,
   useTogglePermission
 } from "../templates/AdminPermissions.hooks";
 import { FetchRolePermissionsOutput } from "../templates/AdminPermissions.db";
-import { Box, Button, Heading, Text } from "@/ui/components";
+import { Box, Button, Heading, Icon, Text } from "@/ui/components";
 
 export const PermissionCard = (props: FetchRolePermissionsOutput) => {
   const {
@@ -109,11 +108,9 @@ export const PermissionCard = (props: FetchRolePermissionsOutput) => {
 
         <Text
           style={{
-            parentWrapper: {
-              textColor: "accent",
-              paddingL: "lg",
-              className: "italic"
-            }
+            textColor: "accent",
+            paddingL: "lg",
+            className: "italic"
           }}
         >
           {description}
