+++ src/features/dashboard/super/authorized-admins/molecules/AdminCard.tsx
@@ -1,9 +1,8 @@
-import { Icon } from "@/tw-styled/ui";
 import { AuthorizedAdmin } from "@prisma/client";
 import { useDeleteAuthorizedAdmin } from "../templates/AuthorizedAdmins.hooks";
 import { useState } from "react";
 import { Modal } from "@/shared/components";
-import { Box, Button } from "@/ui/components";
+import { Box, Button, Icon } from "@/ui/components";
 
 export const AdminCard = ({ admin }: { admin: AuthorizedAdmin }) => {
   /*
