+++ src/features/dashboard/super/authorized-admins/organisms/AdminList.tsx
@@ -1,9 +1,8 @@
-import { Pulse } from "@/tw-styled/ui";
 import { useFetchAuthorizedAdmins } from "../templates/AuthorizedAdmins.hooks";
 import { useIcons } from "@/tw-styled/tools";
 import { useTooltip } from "@/shared/hooks";
 import { AdminCard } from "../molecules";
-import { Box } from "@/ui/components";
+import { Box, Pulse } from "@/ui/components";
 
 export const AdminList = () => {
   const { data, isLoading } = useFetchAuthorizedAdmins();
