+++ src/tw-styled/ui/index.ts
@@ -1,10 +1,5 @@
 export * from "./fieldError/FieldError";
 export * from "./fieldError2/FieldError";
 export * from "./formField/FormField";
-export * from "./icon/Icon";
 export * from "./input/Input";
 export * from "./label/Label";
-export * from "./navbar/Navbar";
-export * from "./pulse/Pulse";
-export * from "./spin/Spin";
-export * from "./submitButton/SubmitButton";
