--- src/tw-styled/ui/pulse/index.ts
@@ -1,1 +1,0 @@
-export * from "./Pulse";
