--- src/tw-styled/ui/spin/Loaders.spin.stories.ts
@@ -1,12 +1,0 @@
-import { Meta, StoryObj } from "@storybook/react";
-import { Spin } from "./Spin";
-
-const meta = {
-  title: "Loaders/Spin",
-  component: Spin
-} satisfies Meta<typeof Spin>;
-
-export default meta;
-export type SpinLoaderStory = StoryObj<typeof meta>;
-
-export const Default: SpinLoaderStory = {};
