+++ src/ui/components/icon/Icon.types.ts
@@ -1,0 +1,8 @@
+import { IconNames, IconVariants } from "@/tw-styled/tools";
+import { BaseStyles } from "@/tw-styled/types";
+
+export type IconProps = {
+  name: IconNames;
+  variant?: IconVariants;
+  style?: BaseStyles;
+};
