+++ src/ui/components/index.ts
@@ -2,5 +2,10 @@ export * from "./box/Box";
 export * from "./button/Button";
 export * from "./form/Form";
 export * from "./heading/Heading";
+export * from "./icon/Icon";
 export * from "./layout/Layout";
+export * from "./navbar/Navbar";
+export * from "./pulse/Pulse";
+export * from "./spin/Spin";
+export * from "./submitButton/SubmitButton";
 export * from "./text/Text";
