+++ src/ui/components/navbar/Navbar.types.ts
@@ -1,0 +1,6 @@
+import { BaseStyles } from "@/tw-styled/types";
+
+export type NavbarProps = {
+  children?: React.ReactNode;
+  style?: BaseStyles;
+};
