+++ src/ui/components/pulse/Pulse.types.ts
@@ -1,0 +1,10 @@
+import { BaseStyles } from "@/tw-styled/types";
+
+export type PulseProps = {
+  size?: "sm" | "md" | "lg";
+  style?: {
+    parentWrapper?: BaseStyles;
+    childrenWrapper?: BaseStyles;
+    children?: BaseStyles;
+  };
+};
