+++ src/ui/components/spin/Spin.types.ts
@@ -1,0 +1,8 @@
+import { BaseStyles } from "@/tw-styled/types";
+
+export type SpinProps = {
+  style?: {
+    parentWrapper?: BaseStyles;
+    icon?: BaseStyles;
+  };
+};
