+++ src/ui/components/submitButton/SubmitButton.types.ts
@@ -1,0 +1,9 @@
+import { BaseStyles, Styles } from "@/tw-styled/types";
+
+export type SubmitButtonProps = {
+  options?: {
+    text?: string;
+    variant?: Styles["buttonVariant"];
+  };
+  style?: BaseStyles;
+};
